Re-org docker api code
