+++ packages/dappmanager/src/calls/fetchDnpRequest.ts
@@ -6,7 +6,7 @@ import shouldUpdate from "../modules/dappGet/utils/shouldUpdate";
 import deepmerge from "deepmerge";
 import { fileToGatewayUrl } from "../utils/distributedFile";
 import { ReleaseFetcher } from "../modules/release";
-import { dockerInfoArchive } from "../modules/docker/dockerApi";
+import { dockerInfoArchive } from "../modules/docker/api";
 import { ComposeEditor, ComposeFileEditor } from "../modules/compose/editor";
 import { parseSpecialPermissions } from "../modules/compose/specialPermissions";
 import {
