+++ packages/dappmanager/src/modules/docker/api/containerInspect.ts
@@ -1,0 +1,13 @@
+import Docker from "dockerode";
+import { docker } from "./docker";
+
+/**
+ * Inspect container
+ * @param containerNameOrId
+ */
+export async function containerInspect(
+  containerNameOrId: string
+): Promise<Docker.ContainerInspectInfo> {
+  const container = docker.getContainer(containerNameOrId);
+  return await container.inspect();
+}
