+++ packages/dappmanager/src/modules/docker/api/docker.ts
@@ -1,0 +1,3 @@
+import Docker from "dockerode";
+
+export const docker = new Docker({ socketPath: "/var/run/docker.sock" });
