+++ packages/dappmanager/src/modules/docker/api/loadImage.ts
@@ -1,0 +1,36 @@
+import { docker } from "./docker";
+
+/**
+ * Load .tar.xz image sending it to the docker daemon
+ */
+export async function loadImage(
+  imagePath: string,
+  onProgress?: (event: DockerLoadProgress) => void
+): Promise<void> {
+  return new Promise((resolve, reject) => {
+    // Must disable quiet flag to receive progress updates
+    docker.loadImage(imagePath, { quiet: "0" }, (err, stream) => {
+      if (err) reject(err);
+      else
+        docker.modem.followProgress(
+          stream,
+          function onFinished(err: Error): void {
+            if (err) reject(err);
+            else resolve();
+          },
+          // eslint-disable-next-line @typescript-eslint/no-empty-function
+          onProgress || ((): void => {})
+        );
+    });
+  });
+}
+
+interface DockerLoadProgress {
+  status: string; // "Loading layer";
+  progressDetail: {
+    current: number; // 221151232;
+    total: number; // 536990720;
+  };
+  progress: string; // "[====================>                              ]  221.2MB/537MB";
+  id: string; // "32c9b213197b";
+}
