+++ packages/dappmanager/src/modules/docker/api/volumeInspect.ts
@@ -1,0 +1,9 @@
+import Docker from "dockerode";
+import { docker } from "./docker";
+
+export function dockerVolumeInspect(
+  volumeName: string
+): Promise<Docker.VolumeInspectInfo> {
+  const volume = docker.getVolume(volumeName);
+  return volume.inspect();
+}
