+++ packages/dappmanager/src/modules/ethClient/migrateEthchain.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import * as db from "../../db";
 import { changeEthMultiClient } from "./changeEthMultiClient";
 import { listContainerNoThrow } from "../docker/list";
-import { dockerVolumesList, dockerDf } from "../docker/dockerApi";
+import { dockerVolumesList, dockerDf } from "../docker/api";
 import { dockerRm, dockerVolumeRm } from "../docker/dockerCommands";
 import { ComposeFileEditor } from "../compose/editor";
 import { migrateVolume } from "../hostScripts";
