+++ packages/dappmanager/src/utils/autoUpdateHelper.ts
@@ -365,7 +365,7 @@ function setPending(dnpName: string, data: AutoUpdatePendingEntry): void {
  * Get an auto-update feedback message
  *
  * @param dnpName "bitcoin.dnp.dappnode.eth"
- * @param currentVersion "0.2.6", must come from dockerList, dnp.version
+ * @param currentVersion "0.2.6", must come from dnp.version
  * @returns feedback = {
  *   updated: 15363818244,
  *   manuallyUpdated: true,
@@ -423,7 +423,7 @@ export function getDnpFeedbackMessage({
  * the logic is different than for a single version DNP
  *
  * @param dnpName "bitcoin.dnp.dappnode.eth"
- * @param currentVersion "0.2.6", must come from dockerList, dnp.version
+ * @param currentVersion "0.2.6", must come from dnp.version
  * @returns feedback = {
  *   updated: 15363818244,
  *   manuallyUpdated: true,
