+++ packages/dappmanager/src/utils/getDappmanagerImage.ts
@@ -1,4 +1,4 @@
-import { dockerList } from "../modules/docker/dockerApi";
+import { listContainers } from "../modules/docker/api";
 import params from "../params";
 import memoize from "memoizee";
 
@@ -6,7 +6,7 @@ const dmName = params.dappmanagerDnpName;
 
 const getDappmanagerImageMemoized = memoize(
   async (): Promise<string> => {
-    const containers = await dockerList({ filters: { name: [dmName] } });
+    const containers = await listContainers({ filters: { name: [dmName] } });
     const dappmanager = containers[0];
 
     if (!dappmanager) throw Error(`No image found for ${dmName}`);
@@ -19,6 +19,6 @@ const getDappmanagerImageMemoized = memoize(
  * Returns the image of the current dappmanager instance running
  * It's memoized since the image will not change until this app is reseted
  */
-export default async function getDappmanagerImage() {
+export default async function getDappmanagerImage(): Promise<string> {
   return getDappmanagerImageMemoized();
 }
